requirements.txt
PyAutoGUI~=0.9.53
Pillow
Requests~=2.26.0
pushbullet.py~=0.12.0
tinydb~=4.5.1